+++ src/main/java/com/wediscussmovies/project/service/ReplyService.java
@@ -5,5 +5,6 @@ import java.util.Optional;
 public interface ReplyService {
     public Reply save(Reply r);
     public void delete(Reply r);
-    Optional<Reply> findById(Integer id);
+    Optional<Reply> findById(Long id);
+
 }
